refactor(clean-arch): split services/ports into protocols and wiring
services/ports.py is now protocols-only (GradeTablesReader/StandardsStore
re-exported from new data/ports/grade_tables.py and
data/ports/standards_store.py). New services/_wiring.py carries the ~90
concretion re-exports plus the additions needed by project_registration,
the report-parser grade helpers, and the standards_store writers. 26
consumer files mechanically retargeted from services.ports to
services._wiring; the 4 Protocol-only consumers stay on services.ports.

New public services/project_index.py facade (shared_repo.py precedent) so
api/import_project.py stops reaching into a services underscore-adjacent
module. New AST guard test asserts services/ports.py never re-imports a
data concretion.

Also: hoist rescore.py's and mutation_rescore.py's deferred imports (no
cycles), delete the dead data/fs/report_parser/_run_lookup.py, reword the
EventLogWriter docstring, add an autouse index-cache-clear fixture, and
drop the _SHARED_RUN_DIM_CACHE/_SHARED_RUN_DIM_LOCK test-only module
aliases in favor of clear_shared_dimension_cache()/DimensionCache.keys().

Zero behavior change except _read_repo_info: a valid-JSON-but-non-dict
repository_info.json now degrades to {} instead of being returned
verbatim (strictly safer).
