+++ ARCHITECTURE.md
@@ -62,7 +62,7 @@ Pre-existing violations are grandfathered in `tools/import_baseline.txt` (a burn
 - Private modules use `_` prefix (e.g., `_fs_projects.py`).
 - Public APIs live in the parent `__init__.py` with re-exports for backward compatibility.
 - Frozen dataclasses for data transfer objects.
-- `services/ports.py` is the single boundary between services and data layers: it re-exports the data functions services use and hosts the Protocols services accept as injected seams (`StandardsStore`, `GradeTablesReader`). `tools/check_imports.py` allows any services→data import — this is a convention, not an enforcement point — but new or edited services code imports through ports.py.
+- `services/ports.py` and `services/_wiring.py` are the single boundary between services and data layers, split by role: `ports.py` hosts only the Protocols services accept as injected seams (`StandardsStore`, `GradeTablesReader`, re-exported from `data/ports/`) and boundary error types; `_wiring.py` is the composition/default-binding module, re-exporting the concrete data-layer functions services use. Services import their types from `ports.py` and their defaults from `_wiring.py`, so a storage swap touches `_wiring` + `data/`, never `ports` consumers. `tools/check_imports.py` allows any services→data import — this is a convention, not an enforcement point — but new or edited services code imports through `_wiring.py` (or `ports.py` for Protocols).
 - Ports for concrete adapters live in `data/ports/` (`FindingsRepository`, `AssistantStore`, `ActionLog` — mirrors `data/actions_log.py::ActionLogWriter`, `DismissedSnippetsReader` — a `Callable` alias mirroring `data/sqlite/findings_queries.py::read_dismissed_snippets`); consumers annotate against the Protocol, composition roots construct the SQLite classes.
 - Entities in `core/` are plain frozen dataclasses (stdlib only). Serialization lives in adapters: the event-log JSON codec is `data/events/codec.py`, dict→dataclass mappers are `data/mappers/`, the API wire serializer (`to_camel_dict`) is `shared/serialization.py`, and env-derived settings are resolved in `config/` (e.g. `config/evidence_env.py`, `config/clone_env.py`, `config/analysis_env.py`, `config/standards_env.py`) and passed inward as parameters. `tests/tools/test_serialization_boundary.py` ratchets `to_camel_dict` use to `api/` plus a declared list of other wire boundaries.
 - Inner layers (`core/`, `analysis/`, `services/`, `config/`) accept an injected `core/observability.py::LogSink` instead of importing a logging framework directly; the default is the silent `NULL_LOG`, and composition roots (CLI, api, dashboard) pass a real sink, e.g. `shared/log_sink.py::SHARED_LOG`. `tests/tools/test_logging_boundary.py` ratchets new `logging`/`shared.logging` imports in those four directories against a declared, burn-down-only list.
