+++ src/quodeq/services/_fs_clone.py
@@ -7,8 +7,7 @@ import logging
 import subprocess as _subprocess
 from pathlib import Path
 
-from quodeq.services.ports import clone_repo
-from quodeq.data.fs.shared_repo import remove_clone_dir
+from quodeq.services._wiring import clone_repo, remove_clone_dir
 from quodeq.config.clone_env import clone_shallow_months, git_clone_timeout_s
 
 _logger = logging.getLogger(__name__)
