+++ src/quodeq/services/_trend_fetcher.py
@@ -30,7 +30,7 @@ from quodeq.services._cache import make_lru_dimension_fetcher
 from quodeq.services.deleted import deleted_keys as _default_deleted_keys
 from quodeq.services.dismissed import dismissed_keys as _default_dismissed_keys
 from quodeq.data.fs.report_parser.runs import read_run_scalars as _default_read_run_scalars
-from quodeq.services.ports import load_suppression_rules
+from quodeq.services._wiring import load_suppression_rules
 from quodeq.services.rescore import _rescore_dimension
 from quodeq.services.score_cache import make_cache_backed_fetcher
 from quodeq.shared.validation import validate_path_segment
