+++ src/quodeq/services/run_events.py
@@ -5,7 +5,7 @@ rules); this facade exposes the Event Log reader through services, mirroring
 how ``services/grade_formula.py`` fronts the params store.
 """
 from quodeq.data.events.reader import EventLogReader  # noqa: F401 — re-exported API
-from quodeq.services.ports import read_dimensions  # noqa: F401,E402 — re-exported API
+from quodeq.services._wiring import read_dimensions  # noqa: F401,E402 — re-exported API
 
 
 def read_run_dim_states(reports_dir, project: str, run_id: str) -> dict:
