+++ src/quodeq/services/suppression.py
@@ -31,7 +31,7 @@ from quodeq.services.suppression_keys import (  # noqa: F401 — re-exported API
     is_dismissed,
 )
 from quodeq.shared.validation import validate_path_segment
-from quodeq.services.ports import (  # noqa: F401 — re-exported API
+from quodeq.services._wiring import (  # noqa: F401 — re-exported API
     load_suppression_rules,
 )
